refactor: move the mongo models around to a more sensible layout
